requirements.txt
boto==2.48.0
boto3==1.7.15
botocore==1.10.15
bz2file==0.98
cachetools==2.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cld2-cffi==0.1.4
click==6.7
cymem==1.31.2
cytoolz==0.8.2
decorator==4.3.0
dill==0.2.7.1
docutils==0.14
en-core-web-sm==2.0.0
feedparser==5.2.1
Flask==1.0.2
ftfy==4.4.3
future==0.16.0
gensim==3.4.0
html5lib==1.0.1
idna==2.6
ijson==2.3
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
networkx==2.1
nltk==3.3
numpy==1.14.2
pathlib==1.0.1
plac==0.9.6
preshed==1.0.0
pycparser==2.18
pyemd==0.5.1
Pyphen==0.9.4
python-dateutil==2.7.2
python-Levenshtein==0.12.0
python-telegram-bot==10.0.2
regex==2017.4.5
requests==2.18.4
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
smart-open==1.5.7
spacy==2.0.11
termcolor==1.1.0
textacy==0.6.1
thinc==6.10.2
toolz==0.9.0
tqdm==4.23.0
ujson==1.35
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
